+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [3.0.4] - 2019-10-08
+
+### Added
+
+- Error handling during generation of iframe.
+- Highlighting of errors by displaying an alert box.
+
+### Fixed
+
+- IFDW\Utils\Host::cleanHost() must be of the type string, null given.
+
 ## [3.0.3] - 2019-09-24
 
 ### Fixed
@@ -200,8 +211,9 @@ All notable changes to this project will be documented in this file.
 
 - Initial release.
 
-[unreleased]: https://github.com/ITS-Boehm/include-fussball-de-widgets/compare/v3.0.3...HEAD
-[3.0.2]: https://github.com/ITS-Boehm/include-fussball-de-widgets/compare/v3.0.2...v3.0.3
+[unreleased]: https://github.com/ITS-Boehm/include-fussball-de-widgets/compare/v3.0.4...HEAD
+[3.0.4]: https://github.com/ITS-Boehm/include-fussball-de-widgets/compare/v3.0.3...v3.0.4
+[3.0.3]: https://github.com/ITS-Boehm/include-fussball-de-widgets/compare/v3.0.2...v3.0.3
 [3.0.2]: https://github.com/ITS-Boehm/include-fussball-de-widgets/compare/v3.0.1...v3.0.2
 [3.0.1]: https://github.com/ITS-Boehm/include-fussball-de-widgets/compare/v3.0.0...v3.0.1
 [3.0.0]: https://github.com/ITS-Boehm/include-fussball-de-widgets/compare/v2.2.2...v3.0.0